Fix movement with mouse on initial load pre-pointer lock. Closes GH-39
Craft disables the GLFW cursor on startup, to allow for mouse look. This
works fine natively, however on the web, pointer lock has to be acquired
by user action -- it cannot be acquired automatically on startup. But
the `glfwSetInputMode(g->window, GLFW_CURSOR, GLFW_CURSOR_DISABLED);`
call attempts to disable it, so the code in handle_mouse_input() thinks
it is disabled (and auto-recentered in the screen), and lets you look
around as if it was, even though it isn't. Clicking to enter pointer
lock, then escaping to exit it, correctly re-enables the GLFW mouse
cursor, working around this problem. To fix the problem on the initial
load, skip disabling the GLFW mouse cursor for emscripten, since it
cannot be.