Fix starting underground caused by incorrect online mode. Closes GH-43
GH-38 added a feature where appending #servername to the document URL
connects to the server automatically, by interpreting the hash string as
arguments for the array argv. This was accomplished by splitting the
string with .split(' '), as a simple first pass (no complicated parsing
to allow command-line arguments with embedded strings quoted), but it
turns out, if you split the empty string, you get back an empty string:
that is, ''.split(' ') is [""], not the empty array []. Since one
argument was passed, argc==2 so g->mode was set to MODE_ONLINE, causing
the client to not properly initialize single player offline mode, in
particular, not setting s->y = highest_block(s->x, s->z) + 2. To fix
this, handle the special case of an empty string meaning no arguments,
correctly allowing single player offline mode to initialize by default.