Copy textures and shaders into build directory. Closes GH-35
Conveniently allows non in-source builds, for the native build which
requires these resources in the same directory it is launched from. With
emscripten and native side-by-side, you can create two build
directories, one for each, and launch ./craft from the native build
directory directly after this fix, since it has the textures/shaders.