Sync glfw cursor to pointer lock disable state. Closes GH-44
Clicking the canvas activates pointer lock, which disables the glfw
mouse cursor, so the game recognizes it as disabled and uses mouse look.
However, pointer lock can be disabled anytime, such as by Cmd-[ to
switch to another browser tab, and back. The glfw cursor would still be
considered enabled, triggering mouselook even without pointer lock. To
fix this, register an emscripten callback to listen for
pointerlockchange events, and re-enable the glfw mouse cursor when the
browser disables pointer lock.