Ignore glfwGetKey() when blurred, partial stick key fix. Closes GH-40
If a key is pressed while the game loses focus (onblur), then it would
be reported as still pressed, appearing stuck, so ignore it. However,
there is still the bug where it will be stuck once the player refocuses
the window - but this is slightly less annoying, so it is an
improvement. Issue reported to emscripten:
https://github.com/kripken/emscripten/issues/5122