Resize window to canvas, using emscripten_get_canvas_size() in one_iter
This changes the game area to match the browser window as it is resized,
when in windowed mode, works in Chrome but since it checks every
iteration, may not be the most efficient.

There exists "canvasResizedCallback" but it is used in
EmscriptenFullscreenStrategy [1] used in emscripten_request_fullscreen_strategy(),
but to enter fullscreen I am using EM_ASM(Module.requestFullscreen(1,1))
since it works in the glfwKeyCallback(). Maybe some other event-driven
mechanism could be used here instead.

[1] https://kripken.github.io/emscripten-site/docs/api_reference/html5.h.html#c.EmscriptenFullscreenStrategy