Connect to /craftws path by default, support ws/wss URLs. Closes GH-60
Previously, due to how the Berkeley sockets API is used and bridged in
Craft through Emscripten, only hostnames and ports could be specified to
connect to servers, which were translated to WebSocket URLs as
ws://hostname:port/ -- that is, limited to ws: scheme and path '/'. By
overriding Module['websocket']['url'], it is now possible to specify
arbitrary ws:// or wss:// URLs in place of the hostname, and with any
/path -- which is now changed to /craftws by default.

`/online localhost` same as `/online ws://localhost:4081/craftws` now.