Fix/workaround click retargeting camera on Firefox. Closes GH-56
For unknown reasons, only on Firefox, there is a mouse movement delta
reported when clicking or releasing the mouse. Followed by a zero
movement delta, as expected, which is all we see on Chrome and Safari.
To fix the movement delta causing unintended rotations and forcing the
player to look downward, ignore the first movement immediately after a
click (press or release). On other browsers, delta is 0, ignored anyway.