Fix WebGL warnings about no bound ARRAY_BUFFER, updates GH-27
Fixes "WebGL: INVALID_OPERATION: vertexAttribPointer: no bound ARRAY_BUFFER"
which was the last error on Safari. The attributes were bound as 0 is
position, 1 is normal, 2 is uv, but most shaders only used position and
uv, so there was a gap between 0 and 2. Reorganize the bindings such
that uv is now 1, so the attribute indices are contigiuous: 0, 1.