Disable the heavy sqlite-based cache for emscripten (GH-107, GH-2)
There are a large number of sqlite calls during startup, when connecting
to an online server, there is additional complexity and page load time,
possibly bugs caused by this heavy dependency (GH-107). This change doesn't
remove it completely (GH-2), it is still used even on the emscripten
port in offline mode (not connected to server), and natively all the
time, but disabling it in online mode in emscripten covers the common
case.