Fix shift crouch sticking with Cmd-Shift-4 screenshots. Closes GH-68
The shift key, used for crouching or descending when flying, would
appear to "stick" if the player presses Cmd-Shift-4 to take a screenshot
when they are flying, because onblur is not called because the window
does not lose focus, it does not lose pointer lock, yet the shift key is
released. This would manifest itself as not being able to fly even
though in fly mode. To fix this, maintain g->crouching and clear it in
on_key() if the Super (Command) key is pressed simultaneously. This
allows pressing Cmd and then Shift to not descend, but Shift by itself
functions normally as before.